Refactor collection types and throw nice errors when there's a problem with them
