--- SUCC/Parsing Logic/Types/Collection Types/CollectionTypes.cs
@@ -1,287 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-
-namespace SUCC.ParsingLogic
-{
-    // This class used to use dynamics instead of this non-generic generics trickery.
-    // However, the old way only worked when compiled with Unity's compiler for Windows.
-    // On other platforms, or when compiled by Visual Studio, they would throw RuntimeBinderExceptions.
-    internal static class CollectionTypes
-    {
-        internal static bool TrySetCollection(Node node, object data, Type collectionType, FileStyle style)
-        {
-            if (collectionType.IsArray) SetArrayNode(node, data, collectionType, style);
-            else if (collectionType.IsList()) SetListNode(node, data, collectionType, style);
-            else if (collectionType.IsHashSet()) SetHashSetNode(node, data, collectionType, style);
-            else if (collectionType.IsDictionary()) SetDictionaryNode(node, data, collectionType, style);
-
-            else if (node.ChildNodeType == NodeChildrenType.list)
-                throw new FormatException($"{collectionType} is not a supported collection type");
-            else return false;
-
-            return true;
-        }
-
-        internal static object TryGetCollection(Node node, Type collectionType)
-        {
-            if (collectionType.IsArray) 
-                return RetrieveArray(node, collectionType);
-
-            if (collectionType.IsList())
-                return RetrieveList(node, collectionType);
-
-            if (collectionType.IsHashSet())
-                return RetrieveHashSet(node, collectionType);
-
-            if (collectionType.IsDictionary())
-                return RetrieveDictionary(node, collectionType);
-
-            if (node.ChildNodeType == NodeChildrenType.list)
-                throw new FormatException($"{collectionType} is not a supported collection type");
-
-            return null;
-        }
-
-        private static bool IsList(this Type type)
-            => type.IsGenericType && type.GetGenericTypeDefinition() == typeof(List<>);
-
-        private static bool IsHashSet(this Type type)
-            => type.IsGenericType && type.GetGenericTypeDefinition() == typeof(HashSet<>);
-
-        private static bool IsDictionary(this Type type)
-            => type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Dictionary<,>);
-
-
-
-        // arrays are the only type in here that doesn't require non-generic generics trickery
-        private static void SetArrayNode(Node node, object array, Type arrayType, FileStyle style)
-        {
-            Type elementType = arrayType.GetElementType();
-
-            var boi = (Array)array;
-
-            node.CapChildCount(boi.Length);
-
-            for (int i = 0; i < boi.Length; i++)
-                NodeManager.SetNodeData(node.GetChildAddressedByListNumber(i), boi.GetValue(i), elementType, style);
-        }
-
-        private static object RetrieveArray(Node node, Type arrayType)
-        {
-            Type elementType = arrayType.GetElementType();
-            var array = Array.CreateInstance(elementType, node.ChildNodes.Count);
-
-            for (int i = 0; i < node.ChildNodes.Count; i++)
-            {
-                ListNode child = node.GetChildAddressedByListNumber(i);
-                object element = NodeManager.GetNodeData(child, elementType);
-                array.SetValue(element, i);
-            }
-
-            return array;
-        }
-
-
-        private static void SetListNode(Node node, object list, Type listType, FileStyle style)
-        {
-            Type elementType = listType.GetGenericArguments()[0];
-
-            SetListNode_.GetBoundGenericMethod(elementType)
-                .Invoke(null, node, list, style);
-        }
-        private static GenericMethodHelper SetListNode_ = new GenericMethodHelper(GetMethod(nameof(SetListNodeGeneric)));
-        private static void SetListNodeGeneric<T>(Node node, List<T> list, FileStyle style)
-        {
-            node.CapChildCount(list.Count);
-
-            for (int i = 0; i < list.Count; i++)
-                NodeManager.SetNodeData<T>(node.GetChildAddressedByListNumber(i), list[i], style);
-        }
-
-        private static object RetrieveList(Node node, Type listType)
-        {
-            Type elementType = listType.GetGenericArguments()[0];
-
-            return RetrieveList_.GetBoundGenericMethod(elementType)
-                .Invoke(null, node);
-        }
-        private static GenericMethodHelper RetrieveList_ = new GenericMethodHelper(GetMethod(nameof(RetrieveListGeneric)));
-        private static List<T> RetrieveListGeneric<T>(Node node)
-        {
-            var list = new List<T>(capacity: node.ChildNodes.Count);
-
-            for (int i = 0; i < node.ChildNodes.Count; i++)
-            {
-                ListNode child = node.GetChildAddressedByListNumber(i);
-                var item = NodeManager.GetNodeData<T>(child);
-                list.Add(item);
-            }
-
-            return list;
-        }
-
-
-        private static void SetHashSetNode(Node node, object hashset, Type hashsetType, FileStyle style)
-        {
-            Type elementType = hashsetType.GetGenericArguments()[0];
-
-            SetHashSetNode_.GetBoundGenericMethod(elementType)
-                .Invoke(null, node, hashset, style);
-        }
-        private static GenericMethodHelper SetHashSetNode_ = new GenericMethodHelper(GetMethod(nameof(SetHashSetNodeGeneric)));
-        private static void SetHashSetNodeGeneric<T>(Node node, HashSet<T> hashset, FileStyle style)
-        {
-            node.CapChildCount(hashset.Count);
-
-            int i = 0;
-            foreach (var item in hashset)
-            {
-                NodeManager.SetNodeData<T>(node.GetChildAddressedByListNumber(i), item, style);
-                i++;
-            }
-        }
-
-        private static object RetrieveHashSet(Node node, Type hashsetType)
-        {
-            Type elementType = hashsetType.GetGenericArguments()[0];
-
-            return RetrieveHashSet_.GetBoundGenericMethod(elementType)
-                .Invoke(null, node);
-        }
-        private static GenericMethodHelper RetrieveHashSet_ = new GenericMethodHelper(GetMethod(nameof(RetrieveHashSetGeneric)));
-        private static HashSet<T> RetrieveHashSetGeneric<T>(Node node)
-        {
-            var hashset = new HashSet<T>(); // todo: use the capacity constructor once unity can use .Net 4.7.2 - i.e. new HashSet<T>(capacity: node.ChildNodes.Count)
-
-            for (int i = 0; i < node.ChildNodes.Count; i++)
-            {
-                ListNode child = node.GetChildAddressedByListNumber(i);
-                var item = NodeManager.GetNodeData<T>(child);
-                hashset.Add(item);
-            }
-
-            return hashset;
-        }
-
-
-        private static void SetDictionaryNode(Node node, object dictionary, Type dictionaryType, FileStyle style, bool forceArrayMode = false)
-        {
-            Type keyType = dictionaryType.GetGenericArguments()[0];
-            Type valueType = dictionaryType.GetGenericArguments()[1];
-
-            SetDictionary_.GetBoundGenericMethod(keyType, valueType)
-                .Invoke(null, node, dictionary, style, forceArrayMode);
-        }
-        private static GenericMethodHelper_2 SetDictionary_ = new GenericMethodHelper_2(GetMethod(nameof(SetDictionaryNodeGeneric)));
-        private static void SetDictionaryNodeGeneric<TKey, TValue>(Node node, Dictionary<TKey, TValue> dictionary, FileStyle style, bool forceArrayMode = false)
-        {
-            bool keyIsBase = BaseTypesManager.IsBaseType(typeof(TKey));
-
-            if (keyIsBase && !forceArrayMode && !style.AlwaysArrayDictionaries)
-            {
-                // we might have switched between standard and array dictionary storage, and if so, children need to be reset
-                if (node.ChildNodeType != NodeChildrenType.key)
-                    node.ClearChildren(newChildrenType: NodeChildrenType.key);
-
-                var CurrentKeys = new List<string>(capacity: dictionary.Count);
-                foreach (var key in dictionary.Keys)
-                {
-                    var value = dictionary[key];
-
-                    string keyAsText = BaseTypesManager.SerializeBaseType<TKey>(key, style);
-
-                    if (!Utilities.IsValidKey(keyAsText))
-                    {
-                        SetDictionaryNodeGeneric(node, dictionary, style, forceArrayMode: true);
-                        return;
-                    }
-
-                    CurrentKeys.Add(keyAsText);
-                    KeyNode child = node.GetChildAddressedByName(keyAsText);
-                    NodeManager.SetNodeData<TValue>(child, value, style);
-                }
-
-                // make sure that old data in the file is deleted when a new dictionary is saved.
-                // node.ClearChildren() is not used because we want to keep comments and whitespace intact as much as possible.
-                foreach (var key in node.GetChildKeys())
-                {
-                    if (!CurrentKeys.Contains(key))
-                        node.RemoveChild(key);
-                }
-            }
-            else // save dictionary as KeyValuePair<TKey, TValue>[]
-            {
-                // we might have switched between standard and array dictionary storage, and if so, children need to be reset
-                if (node.ChildNodeType != NodeChildrenType.list)
-                    node.ClearChildren(newChildrenType: NodeChildrenType.list);
-
-                var array = GetWritableKeyValuePairArray(dictionary);
-                NodeManager.SetNodeData(node, array, array.GetType(), style);
-            }
-        }
-
-        private static object RetrieveDictionary(Node node, Type dictionaryType)
-        {
-            Type keyType = dictionaryType.GetGenericArguments()[0];
-            Type valueType = dictionaryType.GetGenericArguments()[1];
-
-            return RetrieveDictionary_.GetBoundGenericMethod(keyType, valueType)
-                .Invoke(null, node);
-        }
-        private static GenericMethodHelper_2 RetrieveDictionary_ = new GenericMethodHelper_2(GetMethod(nameof(RetrieveDictionaryGeneric)));
-        private static Dictionary<TKey, TValue> RetrieveDictionaryGeneric<TKey, TValue>(Node node)
-        {
-            bool keyIsBase = BaseTypesManager.IsBaseType(typeof(TKey));
-
-            var dictionary = new Dictionary<TKey, TValue>(capacity: node.ChildNodes.Count);
-
-            if (keyIsBase && node.ChildNodeType == NodeChildrenType.key)
-            {
-                foreach (var child in node.ChildNodes)
-                {
-                    string childKey = (child as KeyNode).Key;
-                    var key = BaseTypesManager.ParseBaseType<TKey>(childKey);
-                    var value = NodeManager.GetNodeData<TValue>(child);
-                    dictionary.Add(key, value);
-                }
-            }
-            else
-            {
-                var array = NodeManager.GetNodeData<WritableKeyValuePair<TKey, TValue>[]>(node);
-                foreach (var kvp in array)
-                    dictionary.Add(kvp.key, kvp.value);
-            }
-
-            return dictionary;
-        }
-
-        private static WritableKeyValuePair<TKey, TValue>[] GetWritableKeyValuePairArray<TKey, TValue>(Dictionary<TKey, TValue> boi)
-        {
-            var unwritable = Enumerable.ToArray(boi);
-            var writable = new WritableKeyValuePair<TKey, TValue>[unwritable.Length];
-
-            for (int i = 0; i < unwritable.Length; i++)
-                writable[i] = new WritableKeyValuePair<TKey, TValue>(unwritable[i].Key, unwritable[i].Value);
-
-            return writable;
-        }
-
-        private struct WritableKeyValuePair<TKey, TValue>
-        {
-            public WritableKeyValuePair(TKey key, TValue value)
-            {
-                this.key = key;
-                this.value = value;
-            }
-
-            public TKey key;
-            public TValue value;
-        }
-
-
-        private static MethodInfo GetMethod(string methodName)
-            => typeof(CollectionTypes).GetMethod(methodName, BindingFlags.Static | BindingFlags.NonPublic);
-    }
-}
