+++ SUCC/Parsing Logic/Types/Collection Types/GenericMethodHelpers.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 
 namespace SUCC.ParsingLogic
